[ACP] enabled auth

+++ app/main.py
@@ -55,8 +55,7 @@ def main():
 # ACP
 @app.route('/acp/', methods=['GET'])
 def acp():
-    # if not isAuth():
-    #     return redirect(url_for('acp_login'))
+    if not isAuth(): return redirect(url_for('acp_login'))
     return render_template('acp/main.html')
 
 # Авторизация
@@ -73,6 +72,7 @@ def acp_auth():
 # Магазины
 @app.route('/acp/stores/', methods=['GET'])
 def acp_stores():
+    if not isAuth(): return redirect(url_for('acp_login'))
     sort = request.args.get('sort', 'city')
     stores, sorted_stores_ids = model.getStores(sort)
     managers = model.getManagers()
@@ -80,6 +80,7 @@ def acp_stores():
 
 @app.route('/acp/stores/<path:store_id>', methods=['GET'])
 def acp_store(store_id):
+    if not isAuth(): return redirect(url_for('acp_login'))
     store = model.getStore(store_id)
     manager = model.getManager(store['manager_id'])
     claims = model.getClaims(store_id)
@@ -87,59 +88,70 @@ def acp_store(store_id):
 
 @app.route('/acp/stores/add/', methods=['GET'])
 def acp_store_add():
+    if not isAuth(): return redirect(url_for('acp_login'))
     new_store_id = createStoreId()
     managers = model.getManagers()
     return render_template('acp/add_store.html', managers=managers, new_store_id=new_store_id)
 
 @app.route('/acp/stores/insert', methods=['POST'])
 def acp_store_insert():
+    if not isAuth(): return redirect(url_for('acp_login'))
     model.insertStore(request.form.to_dict())
     return redirect(url_for('acp_stores'))
 
 @app.route('/acp/stores/delete/<path:store_id>', methods=['GET'])
 def acp_store_delete(store_id):
+    if not isAuth(): return redirect(url_for('acp_login'))
     model.deleteStore(store_id)
     return redirect(url_for('acp_stores'))
 
 @app.route('/acp/stores/edit/<path:store_id>', methods=['GET'])
 def acp_store_edit(store_id):
+    if not isAuth(): return redirect(url_for('acp_login'))
     store = model.getStore(store_id)
     managers = model.getManagers()
     return render_template('acp/edit_store.html', store=store, managers=managers)
 
 @app.route('/acp/stores/update', methods=['POST'])
 def acp_store_update():
+    if not isAuth(): return redirect(url_for('acp_login'))
     model.updateStore(request.form.to_dict())
     return redirect(url_for('acp_store_edit', store_id=request.form['id']))
 
 # Менеджеры
 @app.route('/acp/managers/', methods=['GET'])
 def acp_managers():
+    if not isAuth(): return redirect(url_for('acp_login'))
     managers = model.getManagers()
     return render_template('acp/managers.html', managers=managers)
 
 @app.route('/acp/managers/add/', methods=['GET'])
 def acp_manager_add():
+    if not isAuth(): return redirect(url_for('acp_login'))
     managers = model.getManagers()
     return render_template('acp/add_manager.html')
 
 @app.route('/acp/managers/insert', methods=['POST'])
 def acp_manager_insert():
+    if not isAuth(): return redirect(url_for('acp_login'))
     model.insertManager(request.form.to_dict())
     return redirect(url_for('acp_managers'))
 
 @app.route('/acp/managers/delete/<path:manager_id>', methods=['GET'])
 def acp_manager_delete(manager_id):
+    if not isAuth(): return redirect(url_for('acp_login'))
     model.deleteManager(manager_id)
     return redirect(url_for('acp_managers'))
 
 @app.route('/acp/managers/edit/<path:manager_id>', methods=['GET'])
 def acp_manager_edit(manager_id):
+    if not isAuth(): return redirect(url_for('acp_login'))
     managers = model.getManagers()
     return render_template('acp/edit_manager.html', manager=managers[int(manager_id)])
 
 @app.route('/acp/managers/update', methods=['POST'])
 def acp_manager_update():
+    if not isAuth(): return redirect(url_for('acp_login'))
     model.updateМanager(request.form.to_dict())
     return redirect(url_for('acp_manager_edit', manager_id=request.form['id']))
